add find_objects module
