+++ lib/merchant_repository.rb
@@ -1,14 +1,17 @@
 require 'CSV'
 require_relative 'merchant'
+require_relative 'find_module'
 
 class MerchantRepository
+  include FindObjects
   attr_accessor :merchants
 
   def initialize(filepath)
     @merchants = []
     load_merchants(filepath)
+    @storage = @merchants
   end
-  
+
   def inspect
       "#<#{self.class} #{@merchants.size} rows>"
   end
@@ -23,17 +26,17 @@ class MerchantRepository
     end
   end
 
-  def find_by_id(id)
-    @merchants.find do |merchant|
-      merchant.id == id
-    end
-  end
+#  def find_by_id(id)
+#    @merchants.find do |merchant|
+#      merchant.id == id
+#    end
+#  end
 
-  def find_by_name(name)
-    @merchants.find do |merchant|
-      merchant.name.downcase == name.downcase
-    end
-  end
+#  def find_by_name(name)
+#    @merchants.find do |merchant|
+#      merchant.name.downcase == name.downcase
+#    end
+#  end
 
   def find_all_by_name(name)
     @merchants.find_all do |merchant|
@@ -55,8 +58,8 @@ class MerchantRepository
     end
   end
 
-  def delete(id)
-    merchant_to_delete = find_by_id(id)
-    @merchants.delete(merchant_to_delete)
-  end
+#  def delete(id)
+#    merchant_to_delete = find_by_id(id)
+#    @merchants.delete(merchant_to_delete)
+#  end
 end
